+++ RELEASE_NOTES.md
@@ -1,1 +1,2 @@
-
+### Changes
+- Added type, resolution, clock_id and raw_clock_id to manifest.json